Licensing Server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
slowapi==0.1.9
python-multipart==0.0.6
aiohttp==3.9.1
cryptography==42.0.0
